+++ src/main/webapp/WEB-INF/views/user/loginCheck.jsp
@@ -1,7 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <!DOCTYPE html>
-<jsp:useBean id="user" class="user.User" scope="page" />
+<%-- <jsp:useBean id="user" class="user.User" scope="page" />--%>
 <jsp:setProperty name="user" property="userID" />
 <jsp:setProperty name="user" property="userPassword" />
 <html>
